Add database update functions and first webpage
It's now easier to connect to different databases with
db.init(). Also models.py will update both databases when
changes are made to the models (No watch funtion, still
manual execution). To update an external database set a environment
variable that is pointing to that IP address, otherwise it will
default to localhost.